Split egCompute into two files in order to avoid cyclic dependency in egMaterial (material requires egComputeBuffer for buffer handling)
