+++ Horde3D/Source/Horde3DEngine/egComputeBuffer.cpp
@@ -1,4 +1,4 @@
-#include "egCompute.h"
+#include "egComputeBuffer.h"
 #include "egModules.h"
 #include "egRenderer.h"
 #include "egCom.h"
@@ -14,7 +14,7 @@ using namespace std;
 // Compute Buffer Resource
 // =================================================================================================
 
-ComputeBufferResource::ComputeBufferResource( const std::string &name, int flags ) : 
+ComputeBufferResource::ComputeBufferResource( const std::string &name, int flags ) :
 	Resource( ResourceTypes::ComputeBuffer, name, flags & ResourceFlags::NoQuery ),
 	_dataSize( 1024 ), _writeRequested( false ), _bufferID( 0 ), _data( nullptr ), _mapped( false ), _numElements( 0 ), _useAsVertexBuf( false ),
 	_vertexLayout( 0 ), _drawType( -1 ), _geoID( 0 ), _geometryParamsSet( false )
@@ -58,7 +58,7 @@ void ComputeBufferResource::release()
 
 bool ComputeBufferResource::load( const char *data, int size )
 {
-//	if ( !Resource::load( data, size ) ) return false;
+	//	if ( !Resource::load( data, size ) ) return false;
 
 	// currently not implemented
 
@@ -173,7 +173,7 @@ int ComputeBufferResource::getElemParamI( int elem, int elemIdx, int param ) con
 
 			break;
 		}
-		
+
 		case ComputeBufferResData::DrawParamsElem:
 		{
 			switch ( param )
@@ -206,7 +206,7 @@ void ComputeBufferResource::setElemParamI( int elem, int elemIdx, int param, int
 						initDefault();
 					}
 					else _dataSize = value;
-	
+
 					return;
 
 				case ComputeBufferResData::CompBufUseAsVertexBufferI:
@@ -385,158 +385,5 @@ void ComputeBufferResource::unmapStream()
 	_mapped = false;
 }
 
-
-// =================================================================================================
-// Compute Node
-// =================================================================================================
-
-ComputeNode::ComputeNode( const ComputeNodeTpl &computeTpl ) : SceneNode( computeTpl )
-{
-	_compBufferRes = computeTpl.compBufRes;
-	_materialRes = computeTpl.matRes;
-
-	_renderable = true;
-
-	_localBBox.min = Vec3f( 0, 0, 0 );
-	_localBBox.max = Vec3f( 1, 1, 1 );
-}
-
-
-ComputeNode::~ComputeNode()
-{
-
-}
-
-
-SceneNodeTpl *ComputeNode::parsingFunc( map< string, string > &attribs )
-{
-// 	map< string, string >::iterator itr;
-// 	ComputeNodeTpl *computeTpl = new ComputeNodeTpl( "", nullptr, nullptr );
-
-	return nullptr;
-}
-
-
-SceneNode *ComputeNode::factoryFunc( const SceneNodeTpl &nodeTpl )
-{
-	if ( nodeTpl.type != SceneNodeTypes::Compute ) return nullptr;
-
-	return new ComputeNode( *( ComputeNodeTpl * ) &nodeTpl );
-}
-
-
-void ComputeNode::onPostUpdate()
-{
-	_bBox = _localBBox;
-	_bBox.transform( _absTrans );
-}
-
-
-int ComputeNode::getParamI( int param ) const
-{
-	switch ( param )
-	{
-		case ComputeNodeParams::CompBufResI:
-			if ( _compBufferRes ) return _compBufferRes->getHandle();
-			else return 0;
-		case ComputeNodeParams::MatResI:
-			if ( _materialRes ) return _materialRes->getHandle();
-			else return 0;
-		default:
-			break;
-	}
-
-	return SceneNode::getParamI( param );
-}
-
-
-void ComputeNode::setParamI( int param, int value )
-{
-	Resource *res;
-	
-	switch ( param )
-	{
-		case ComputeNodeParams::CompBufResI:
-			res = Modules::resMan().resolveResHandle( value );
-			if ( res == 0x0 || res->getType() == ResourceTypes::ComputeBuffer )
-				_compBufferRes = ( ComputeBufferResource * ) res;
-			else
-				Modules::setError( "Invalid handle in h3dSetNodeParamI for H3DComputeNode::CompBufResI" );
-			return;
-		case ComputeNodeParams::MatResI:
-			res = Modules::resMan().resolveResHandle( value );
-			if ( res == 0x0 || res->getType() == ResourceTypes::Material )
-				_materialRes = ( MaterialResource * ) res;
-			else
-				Modules::setError( "Invalid handle in h3dSetNodeParamI for H3DComputeNode::MatResI" );
-			return;
-		default:
-			break;
-	}
-
-	SceneNode::setParamI( param, value );
-}
-
-
-float ComputeNode::getParamF( int param, int compIdx ) const
-{
-	switch ( param )
-	{
-		case ComputeNodeParams::AABBMinF:
-			if ( compIdx < 0 || compIdx > 2 ) 
-			{
-				Modules::setError( "Invalid compIdx specified in h3dGetNodeParamF for H3DComputeNode::AABBMinF" );
-				return Math::NaN;
-			}
-			
-			return _localBBox.min[ compIdx ];
-		case ComputeNodeParams::AABBMaxF:
-			if ( compIdx < 0 || compIdx > 2 )
-			{
-				Modules::setError( "Invalid compIdx specified in h3dGetNodeParamF for H3DComputeNode::AABBMaxF" );
-				return Math::NaN;
-			}
-
-			return _localBBox.max[ compIdx ];
-		default:
-			break;
-	}
-
-	return SceneNode::getParamF( param, compIdx );
-}
-
-
-void ComputeNode::setParamF( int param, int compIdx, float value )
-{
-	switch ( param )
-	{
-		case ComputeNodeParams::AABBMinF:
-			if ( compIdx < 0 || compIdx > 2 )
-			{
-				Modules::setError( "Invalid compIdx specified in h3dSetNodeParamF for H3DComputeNode::AABBMinF" );
-				return;
-			}
-
-			_localBBox.min[ compIdx ] = value;
-			markDirty();
-	
-			return;
-		case ComputeNodeParams::AABBMaxF:
-			if ( compIdx < 0 || compIdx > 2 )
-			{
-				Modules::setError( "Invalid compIdx specified in h3dSetNodeParamF for H3DComputeNode::AABBMaxF" );
-				return;
-			}
-
-			_localBBox.max[ compIdx ] = value;
-			markDirty();
-
-			return;
-		default:
-			break;
-	}
-	
-	SceneNode::setParamF( param, compIdx, value );
-}
-
 } // namespace
+
